backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.15.0,<4.0
djangorestframework-simplejwt>=5.3.0,<6.0
django-cors-headers>=4.0.0,<5.0
psycopg2-binary>=2.9.0,<3.0
requests>=2.28.0,<3.0
